PIXI-161, PIXI-162, PIXI-231 (#3)
Major rewrite of the hotel splitter container. This is intended to
complete PIXI-161 by co-registering the CT and PET scan cuts to be
correctly overlaying each other, as well as PIXI-162 which deals with
failing runs of the splitter due to difficult data. During the course of
this work, I found that a greater rewrite was needed to ensure that the
container is able to perform high quality splits at a high rate of
success. I will outline some of the changes below:

- Removing the harmonization system as well as the input size field,
both of which, up until now, forced all cuts to be the same size. Due to
the breadth of possible mouse sizes, we can't allow for cuts featuring
small mice to be sized the same as those of large mice as this tends to
cause overlaps -- where one cut shows parts of a second, unintended
mouse.
- New checking for incompatible number of detected mice, e.g. input
expects 4 mice to be found, but the CT splitter finds 5 regions. In some
cases the new system will attempt to merge regions in order to find the
correct number of cuts to make. In others, the newly created
co-registration system allows us to default to using the (more accurate)
regions found by the other modality.
- Alongside the co-registration system, I have added an expansion system
to allow some of the work of translating the PET and CT cuts so that
they overlay each other to be performed by expanding the cuts rather
than moving them. This seems to help with accuracy. I have put a cap on
the amount of allowable expansion to avoid the aforementioned
overlapping cut problem.
- Adding a system to handle multiple sets of PET and CT scans within the
same session. For this, I have chosen use the scan time to connect scans
which occur most closely together and assume they should be the ones
co-registered. To do this, I have added a set of API calls to obtain the
scan time for each scan in cases of more than one scan per modality.
- Removing any previous QC snapshots from previous runs of the hotel
splitter prior to uploading new snapshots which running the splitter a
second time on the same session. This will make showing these snapshots
to the user much easier. This requirement has been outlined in PIXI-231.
- Adding a margin field which allows a user to input a value for the
margin around a region that has been identified for splitting
(ostensibly this is the region featuring a mouse). This system will
default to being unused -- and thus reverting to the hard coded margin
used for all other splits -- but will hopefully provide users with a
method to perform a second pass on a particularly frustrating mouse
configuration that results in an improper split. Basically, this is an
admission that no margin is always going to work, and that, while
several different people have tried to set an appropriate margin for
most images, it may need to be tweaked in certain cases.
- Ensuring that the margin (whether that be the base margin or the user
input margin) is used for every possible type of input scan. Before,
only splits of less than 3 mice added the margin -- likely due to it
creating a higher risk of overlapping cuts. To reduce this worry, I have
reduced the size of the margin in these cases relative to what it is for
2 mice splits.
- A significant amount of general cleanup: removing several methods
which were unused, or which never returned anything due to what was
input by the container; removing outdated method calls that had been
commented out previously; removing unnecessary logging statements to try
to make the output more streamlined.

This is a large PR, but I can't think of any way that it could be
meaningfully split up, as all of changes impact each other. Please let
me know if you have any questions.